requirements.txt
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
PyGetWindow==0.0.9
pyobjc-core==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Quartz==7.1
PyRect==0.1.4
pytesseract==0.3.7
